# target/classes/swagger/swagger-input.yml
swagger: "2.0"
info:
  description: "Manage business"
  version: "1.0.0"
  title: "Business"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "petstore.swagger.io"
basePath: "/v2"
tags:
- name: "Business"
  description: "Everything about business"
schemes:
- "https"
- "http"
paths:
  /v1/user:
    post:
      tags:
      - "user"
      summary: "Add a new user"
      description: "add new user in to database"
      operationId: "addUser"
      consumes:
      - "application/json"
      produces:
      - "text/plain"
      parameters:
      - in: "body"
        name: "body"
        description: "user object need to be add in database"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "successfully created"
          schema:
            type: string
            description: ID of the user being created.
        405:
          description: "Invalid input"
        500:
          description: Internal server error
    get:
      tags:
      - "user"
      summary: "get all user"
      description: "get all user"
      operationId: "getAllUser"
      produces:
      - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/User"
        400:
          description: "Invalid status value"
  /v1/user/{id}:
    put:
      tags:
      - "user"
      summary: "Update an existing user"
      description: "just update the user by id"
      operationId: "updateUser"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "id"
        description: "ID of user to update"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "user object need to be updated"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "updated successfully"
          schema:
            type: "string"
        400:
          description: "Invalid ID supplied"
        404:
          description: "user not found"
        405:
          description: "Validation exception"
    get:
      tags:
      - "user"
      summary: "get user by id"
      description: "get user by id from path"
      operationId: "searchUserByID"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "id"
        description: "ID of user to get"
        required: true
        type: "string"
      responses:
        200:
          description: "successfully updated"
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Invalid ID supplied"
        404:
          description: "user not found"
        405:
          description: "Validation exception"
    delete:
      tags:
      - "user"
      summary: "delete user by id"
      description: "delete user by id from path"
      operationId: "deleteUser"
      parameters:
      - in: "path"
        name: "id"
        description: "ID of user to delete"
        required: true
        type: "string"
      responses:
        200:
          description: "successfully delete"
          schema:
            type: "string"
        400:
          description: "Invalid ID supplied"
        404:
          description: "user not found"
        405:
          description: "Validation exception"
  /v1/user/:
    get:
      tags:
      - "user"
      summary: "get user by name"
      description: "get user by name from path"
      operationId: "searchUserByName"
      produces:
      - "application/json"
      parameters:
      - name: "firstName"
        in: "query"
        description: "firstname of user"
        required: true
        type: "string"
      - name: "lastName"
        in: "query"
        description: "last name of user"
        required: true
        type: "string"
      responses:
        200:
          description: "successfully updated"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/User"
        400:
          description: "Invalid name supplied"
        404:
          description: "User not found"
        405:
          description: "Validation exception"
  /v1/{userId}/user/activate:
    put:
      tags:
      - "admin"
      summary: Activate user.
      description:  Activate user .
      operationId: updateActivateUser
      produces:
      - application/json
      parameters:
      - in: path
        name: userId
        type: string
        description: user Id
        required: true
      responses:
        '200':
          description: "successfull update"
          schema:
            type: string
            description: activate user.
        '500':
          description: Internal Server Error
  /v1/{userId}/user/deactivate:
    put:
      tags:
      - "admin"
      summary: Deactivate user.
      description:  Deactivate user .
      operationId: updateDeactivateUser
      produces:
      - application/json
      parameters:
      - in: path
        name: userId
        type: string
        description: user Id
        required: true
      responses:
        '200':
          description: "successfull update"
          schema:
            type: string
            description: Deactivate user.
        '500':
          description: Internal Server Error
  /v1/{userId}/setLimited:
    put:
      tags:
      - "user"
      summary: set the monthly limited expenses for notification.
      description: set the monthly limited expenses for notification .
      operationId: setLimitedMonthly
      produces:
      - application/json
      parameters:
      - in: path
        name: userId
        type: string
        description: user Id
        required: true
      - in: query
        name: limit
        type: string
        description: monthly limited
        required: true
      responses:
        '200':
          description: "set the monthly limit"
          schema:
            type: string
            description: set the monthly limit.
        '500':
          description: Internal Server Error
  /v1/invoice/{userId}:
    post:
      tags:
      - "invoice"
      summary: add a new invoice  .
      description: add a new invoice  .
      operationId: addInvoice
      produces:
      - text/plain
      parameters:
      - in: path
        name: userId
        type: string
        description: Id of user
        required: true
      - in: "body"
        name: "body"
        description: "Invoice object need to be add in Invoice"
        required: true
        schema:
          $ref: "#/definitions/Invoice"
      responses:
        '200':
          description: "successfull created"
          schema:
            type: string
            description: ID of the invoice being created.
        '400':
          description: Bad request, validation error
        '404':
          description: The invoice is not found.
        '500':
          description: Internal Server Error
  /v1/{userId}/invoices/{invoiceNo}:
    put:
      tags:
      - "invoice"
      summary: Update an existing invoice
      description: just update the invoice by invoiceNo
      operationId: updateInvoice
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: path
        name: invoiceNo
        description: ID of invoice to update
        required: true
        type: string
      - in: path
        name: userId
        description: ID of user exist
        required: true
        type: string
      - in: body
        name: body
        description: "invoice object need to be updated"
        required: true
        schema:
          $ref: "#/definitions/Invoice"
      responses:
        200:
          description: "successfull update"
          schema:
            type: string
        400:
          description: "Invalid ID supplied"
        404:
          description: "Invoice not found"
        405:
          description: "Validation exception"
    delete:
      tags:
      - "invoice"
      summary: delete invoice by id
      description: delete invoice by id from path
      operationId: deleteInvoice
      parameters:
      - in: path
        name: invoiceNo
        description: ID of invoice to delete
        required: true
        type: string
      - in: path
        name: userId
        description: ID of user exist
        required: true
        type: string
      responses:
        200:
          description: "successfull delete"
          schema:
            type: string
        400:
          description: "Invalid ID supplied"
        404:
          description: "Invoice not found"
        405:
          description: "Validation exception"
  /v1/invoices/{invoiceNo}:
    get:
      tags:
      - "invoice"
      summary: Get the invoice by  invoiceNo.
      description: Get the invoice by invoiceNo.
      operationId: readInvoice
      produces:
      - application/json
      - text/plain
      parameters:
      - in: path
        name: invoiceNo
        type: string
        description: invoice code
        required: true
      responses:
        '200':
          description: Invoice specified by Invoice No.
          schema:
            $ref: "#/definitions/Invoice"
        '404':
          description: The invoice is not found.
        '500':
          description: Internal Server Error
  /v1/{userId}/invoice/lists:
    get:
      tags:
      - "invoice"
      summary: Get the list of invoices.
      description: Get the list of invoices by userId.
      operationId: readAllInvoice
      produces:
      - application/json
      parameters:
      - in: path
        name: userId
        type: string
        description: user Id
        required: true
      responses:
        '200':
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Invoice"
        '500':
          description: Internal Server Error
  /v1/{userId}/invoice/viewReport:
    get:
      tags:
        - "invoice"
      summary: Report invoice monthly / yearly.
      description:  Report invoice monthly / yearly by userId.
      operationId: viewReportInvoice
      produces:
        - application/json
      parameters:
        - in: path
          name: userId
          type: string
          description: user Id
          required: true
        - in: query
          name: period
          type: string
          description: changed period .
          required: true
        - in: query
          name: monthly
          description: "input month or year need view report"
          required: true
          type: string
      responses:
        '200':
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Invoice"
        '500':
          description: Internal Server Error
  /v1/{userId}/sendEmail:
    get:
      tags:
        - "mail"
      summary: "send mail"
      description: "send mail"
      operationId: "sendMail"
      produces:
        - "application/json"
      parameters:
      - in: path
        name: userId
        type: string
        description: user Id
        required: true
      - in: query
        name: mailTo
        type: string
        description: Mail To .
        required: false
      - in: query
        name: mailSubject
        type: string
        description: Mail Subject.
        required: false
      - in: query
        name: mailText
        type: string
        description: Mail Text.
        required: false
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
        400:
          description: "Invalid status value"
definitions:
  Invoice:
    type: "object"
    properties:
      typeOfInvoice:
        type: "string"
        example: "Electric"
      amountOfMoney:
        type: "integer"
        example: "300000"
      vat:
        type: "string"
        example: "0.5"
      period:
        type: "string"
        example: "monthly"
      invoiceNo:
        type: "string"
        example: "819d15a3-9692-19ba-9974-180df0e334c1"
      userId:
        type: "string"
        example: "001"
  User:
    type: "object"
    properties:
      id:
        type: "string"
        example: "001"
      firstName:
        type: "string"
        example: "Nhan"
      lastName:
        type: "string"
        example: "Phan"
      email:
        type: "string"
        example: "nhanzks165@gmail.com"
      address:
        type: "string"
        example: "70 lth gv"
      activated:
        readOnly: true
        type: boolean
        example: false
      limited:
        readOnly: true
        type: integer
        example: 0