# target/generated-sources/src/main/java/com/dxc/business/api/V1Api.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.3.1).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.dxc.business.api;

import com.dxc.business.api.model.Invoice;
import com.dxc.business.api.model.User;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;

@Api(value = "v1", description = "the v1 API")
public interface V1Api {

    @ApiOperation(value = "add a new invoice  .", nickname = "addInvoice", notes = "add a new invoice  .", response = String.class, tags={ "invoice", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successfull created", response = String.class),
        @ApiResponse(code = 400, message = "Bad request, validation error"),
        @ApiResponse(code = 404, message = "The invoice is not found."),
        @ApiResponse(code = 500, message = "Internal Server Error") })
    @RequestMapping(value = "/v1/invoice/{userId}",
        produces = { "text/plain" }, 
        method = RequestMethod.POST)
    ResponseEntity<String> addInvoice(@ApiParam(value = "Id of user",required=true) @PathVariable("userId") String userId,@ApiParam(value = "Invoice object need to be add in Invoice" ,required=true )  @Valid @RequestBody Invoice body);


    @ApiOperation(value = "Add a new user", nickname = "addUser", notes = "add new user in to database", response = String.class, tags={ "user", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successfully created", response = String.class),
        @ApiResponse(code = 405, message = "Invalid input"),
        @ApiResponse(code = 500, message = "Internal server error") })
    @RequestMapping(value = "/v1/user",
        produces = { "text/plain" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    ResponseEntity<String> addUser(@ApiParam(value = "user object need to be add in database" ,required=true )  @Valid @RequestBody User body);


    @ApiOperation(value = "delete invoice by id", nickname = "deleteInvoice", notes = "delete invoice by id from path", response = String.class, tags={ "invoice", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successfull delete", response = String.class),
        @ApiResponse(code = 400, message = "Invalid ID supplied"),
        @ApiResponse(code = 404, message = "Invoice not found"),
        @ApiResponse(code = 405, message = "Validation exception") })
    @RequestMapping(value = "/v1/{userId}/invoices/{invoiceNo}",
        method = RequestMethod.DELETE)
    ResponseEntity<String> deleteInvoice(@ApiParam(value = "ID of invoice to delete",required=true) @PathVariable("invoiceNo") String invoiceNo,@ApiParam(value = "ID of user exist",required=true) @PathVariable("userId") String userId);


    @ApiOperation(value = "delete user by id", nickname = "deleteUser", notes = "delete user by id from path", response = String.class, tags={ "user", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successfully delete", response = String.class),
        @ApiResponse(code = 400, message = "Invalid ID supplied"),
        @ApiResponse(code = 404, message = "user not found"),
        @ApiResponse(code = 405, message = "Validation exception") })
    @RequestMapping(value = "/v1/user/{id}",
        method = RequestMethod.DELETE)
    ResponseEntity<String> deleteUser(@ApiParam(value = "ID of user to delete",required=true) @PathVariable("id") String id);


    @ApiOperation(value = "get all user", nickname = "getAllUser", notes = "get all user", response = User.class, responseContainer = "List", tags={ "user", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = User.class, responseContainer = "List"),
        @ApiResponse(code = 400, message = "Invalid status value") })
    @RequestMapping(value = "/v1/user",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<List<User>> getAllUser();


    @ApiOperation(value = "Get the list of invoices.", nickname = "readAllInvoice", notes = "Get the list of invoices by userId.", response = Invoice.class, responseContainer = "List", tags={ "invoice", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = Invoice.class, responseContainer = "List"),
        @ApiResponse(code = 500, message = "Internal Server Error") })
    @RequestMapping(value = "/v1/{userId}/invoice/lists",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<List<Invoice>> readAllInvoice(@ApiParam(value = "user Id",required=true) @PathVariable("userId") String userId);


    @ApiOperation(value = "Get the invoice by  invoiceNo.", nickname = "readInvoice", notes = "Get the invoice by invoiceNo.", response = Invoice.class, tags={ "invoice", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Invoice specified by Invoice No.", response = Invoice.class),
        @ApiResponse(code = 404, message = "The invoice is not found."),
        @ApiResponse(code = 500, message = "Internal Server Error") })
    @RequestMapping(value = "/v1/invoices/{invoiceNo}",
        produces = { "application/json", "text/plain" }, 
        method = RequestMethod.GET)
    ResponseEntity<Invoice> readInvoice(@ApiParam(value = "invoice code",required=true) @PathVariable("invoiceNo") String invoiceNo);


    @ApiOperation(value = "get user by id", nickname = "searchUserByID", notes = "get user by id from path", response = User.class, tags={ "user", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successfully updated", response = User.class),
        @ApiResponse(code = 400, message = "Invalid ID supplied"),
        @ApiResponse(code = 404, message = "user not found"),
        @ApiResponse(code = 405, message = "Validation exception") })
    @RequestMapping(value = "/v1/user/{id}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<User> searchUserByID(@ApiParam(value = "ID of user to get",required=true) @PathVariable("id") String id);


    @ApiOperation(value = "get user by name", nickname = "searchUserByName", notes = "get user by name from path", response = User.class, responseContainer = "List", tags={ "user", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successfully updated", response = User.class, responseContainer = "List"),
        @ApiResponse(code = 400, message = "Invalid name supplied"),
        @ApiResponse(code = 404, message = "User not found"),
        @ApiResponse(code = 405, message = "Validation exception") })
    @RequestMapping(value = "/v1/user/",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<List<User>> searchUserByName(@NotNull @ApiParam(value = "firstname of user", required = true) @Valid @RequestParam(value = "firstName", required = true) String firstName,@NotNull @ApiParam(value = "last name of user", required = true) @Valid @RequestParam(value = "lastName", required = true) String lastName);


    @ApiOperation(value = "send mail", nickname = "sendMail", notes = "send mail", response = String.class, tags={ "mail", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = String.class),
        @ApiResponse(code = 400, message = "Invalid status value") })
    @RequestMapping(value = "/v1/{userId}/sendEmail",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<String> sendMail(@ApiParam(value = "user Id",required=true) @PathVariable("userId") String userId,@ApiParam(value = "Mail To .") @Valid @RequestParam(value = "mailTo", required = false) String mailTo,@ApiParam(value = "Mail Subject.") @Valid @RequestParam(value = "mailSubject", required = false) String mailSubject,@ApiParam(value = "Mail Text.") @Valid @RequestParam(value = "mailText", required = false) String mailText);


    @ApiOperation(value = "set the monthly limited expenses for notification.", nickname = "setLimitedMonthly", notes = "set the monthly limited expenses for notification .", response = String.class, tags={ "user", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "set the monthly limit", response = String.class),
        @ApiResponse(code = 500, message = "Internal Server Error") })
    @RequestMapping(value = "/v1/{userId}/setLimited",
        produces = { "application/json" }, 
        method = RequestMethod.PUT)
    ResponseEntity<String> setLimitedMonthly(@ApiParam(value = "user Id",required=true) @PathVariable("userId") String userId,@NotNull @ApiParam(value = "monthly limited", required = true) @Valid @RequestParam(value = "limit", required = true) String limit);


    @ApiOperation(value = "Activate user.", nickname = "updateActivateUser", notes = "Activate user .", response = String.class, tags={ "admin", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successfull update", response = String.class),
        @ApiResponse(code = 500, message = "Internal Server Error") })
    @RequestMapping(value = "/v1/{userId}/user/activate",
        produces = { "application/json" }, 
        method = RequestMethod.PUT)
    ResponseEntity<String> updateActivateUser(@ApiParam(value = "user Id",required=true) @PathVariable("userId") String userId);


    @ApiOperation(value = "Deactivate user.", nickname = "updateDeactivateUser", notes = "Deactivate user .", response = String.class, tags={ "admin", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successfull update", response = String.class),
        @ApiResponse(code = 500, message = "Internal Server Error") })
    @RequestMapping(value = "/v1/{userId}/user/deactivate",
        produces = { "application/json" }, 
        method = RequestMethod.PUT)
    ResponseEntity<String> updateDeactivateUser(@ApiParam(value = "user Id",required=true) @PathVariable("userId") String userId);


    @ApiOperation(value = "Update an existing invoice", nickname = "updateInvoice", notes = "just update the invoice by invoiceNo", response = String.class, tags={ "invoice", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successfull update", response = String.class),
        @ApiResponse(code = 400, message = "Invalid ID supplied"),
        @ApiResponse(code = 404, message = "Invoice not found"),
        @ApiResponse(code = 405, message = "Validation exception") })
    @RequestMapping(value = "/v1/{userId}/invoices/{invoiceNo}",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.PUT)
    ResponseEntity<String> updateInvoice(@ApiParam(value = "ID of invoice to update",required=true) @PathVariable("invoiceNo") String invoiceNo,@ApiParam(value = "ID of user exist",required=true) @PathVariable("userId") String userId,@ApiParam(value = "invoice object need to be updated" ,required=true )  @Valid @RequestBody Invoice body);


    @ApiOperation(value = "Update an existing user", nickname = "updateUser", notes = "just update the user by id", response = String.class, tags={ "user", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "updated successfully", response = String.class),
        @ApiResponse(code = 400, message = "Invalid ID supplied"),
        @ApiResponse(code = 404, message = "user not found"),
        @ApiResponse(code = 405, message = "Validation exception") })
    @RequestMapping(value = "/v1/user/{id}",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.PUT)
    ResponseEntity<String> updateUser(@ApiParam(value = "ID of user to update",required=true) @PathVariable("id") String id,@ApiParam(value = "user object need to be updated" ,required=true )  @Valid @RequestBody User body);


    @ApiOperation(value = "Report invoice monthly / yearly.", nickname = "viewReportInvoice", notes = "Report invoice monthly / yearly by userId.", response = Invoice.class, responseContainer = "List", tags={ "invoice", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = Invoice.class, responseContainer = "List"),
        @ApiResponse(code = 500, message = "Internal Server Error") })
    @RequestMapping(value = "/v1/{userId}/invoice/viewReport",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<List<Invoice>> viewReportInvoice(@ApiParam(value = "user Id",required=true) @PathVariable("userId") String userId,@NotNull @ApiParam(value = "changed period .", required = true) @Valid @RequestParam(value = "period", required = true) String period,@NotNull @ApiParam(value = "input month or year need view report", required = true) @Valid @RequestParam(value = "monthly", required = true) String monthly);

}